+++ CHANGES.md
@@ -10,7 +10,7 @@
 
   - <https://github.com/georust/gdal/pull/504>
 
-- Added `DriverManager::guess_drivers_for_write` for the ability to auto detect compatible `Driver`s for writing data.
+- Added `DriverManager::guess_driver_for_write` and `DriverManager::guess_drivers_for_write` for the ability to auto detect compatible `Driver`(s) for writing data.
 
 - Added `Feature::unset_field`